graphs/iso: Add solve_old.go for code that is not yet tested.
