+++ iso/candidate.go
@@ -83,47 +83,9 @@ func isPotential(g, s *dot.Node, sub *graphs.SubGraph) bool {
 	if s.Name != sub.Entry() && len(g.Preds) != len(s.Preds) {
 		return false
 	}
-
 	// Verify successors.
 	if s.Name != sub.Exit() && len(g.Succs) != len(s.Succs) {
 		return false
 	}
-
 	return true
-}
-
-// SetPair marks the given node pair as known by removing it from c and storing
-// it in m. As the graph node name is no longer a valid candidate it is removed
-// from all other node pairs in c.
-func (eq *Equation) SetPair(sname, gname string) error {
-	// Sanity check.
-	if key, ok := getKey(eq.m, gname); ok {
-		return errutil.Newf("invalid mapping; sub node %q and %q both map to graph node %q", key, sname, gname)
-	}
-
-	// Move unique node pair from c to m.
-	eq.m[sname] = gname
-	delete(eq.c, sname)
-
-	// Remove graph node name of the unique node pair from all other node
-	// pairs in c.
-	for key, candidates := range eq.c {
-		delete(candidates, gname)
-		if len(eq.c[key]) == 0 {
-			return errutil.Newf("invalid mapping; sub node %q has no candidates", key)
-		}
-	}
-
-	return nil
-}
-
-// getKey returns the first key in m which maps to the value val. The boolean
-// value is true if such a key could be located, and false otherwise.
-func getKey(m map[string]string, val string) (key string, ok bool) {
-	for key, x := range m {
-		if x == val {
-			return key, true
-		}
-	}
-	return "", false
 }
